+++ buildpacks/ruby/src/main.rs
@@ -2,9 +2,6 @@ use bullet_stream::{style, Print};
 use commons::cache::CacheError;
 use commons::gemfile_lock::GemfileLock;
 use commons::metadata_digest::MetadataDigest;
-#[allow(clippy::wildcard_imports)]
-use commons::output::build_log::*;
-use commons::output::warn_later::WarnGuard;
 use core::str::FromStr;
 use fs_err::PathExt;
 use fun_run::CmdError;
@@ -111,11 +108,8 @@ impl Buildpack for RubyBuildpack {
     }
 
     #[allow(clippy::too_many_lines)]
-    #[allow(deprecated)]
     fn build(&self, context: BuildContext<Self>) -> libcnb::Result<BuildResult, Self::Error> {
         let mut build_output = Print::new(stdout()).h2("Heroku Ruby Buildpack");
-        let logger = BuildLog::new(stdout()).without_buildpack_name();
-        let warn_later = WarnGuard::new(stdout());
 
         // ## Set default environment
         let (mut env, store) =
@@ -185,7 +179,7 @@ impl Buildpack for RubyBuildpack {
         };
 
         // ## Bundle install
-        (_, env) = {
+        (build_output, env) = {
             let bullet = build_output.bullet("Bundle install gems");
             let (bullet, layer_env) = layers::bundle_install_layer::handle(
                 &context,
@@ -219,30 +213,33 @@ impl Buildpack for RubyBuildpack {
         };
 
         // ## Detect gems
-        let (mut logger, gem_list, default_process) = {
-            let section = logger.section("Setting default processes");
+        let (mut build_output, gem_list, default_process) = {
+            let bullet = build_output.bullet("Default process detection");
 
-            let gem_list = gem_list::GemList::from_bundle_list(&env, section.as_ref())
-                .map_err(RubyBuildpackError::GemListGetError)?;
-            let default_process = steps::get_default_process(section.as_ref(), &context, &gem_list);
+            let (bullet, gem_list) =
+                gem_list::bundle_list(bullet, &env).map_err(RubyBuildpackError::GemListGetError)?;
+            let (bullet, default_process) = steps::get_default_process(bullet, &context, &gem_list);
 
-            (section.end_section(), gem_list, default_process)
+            (bullet.done(), gem_list, default_process)
         };
 
         // ## Assets install
-        logger = {
-            let section = logger.section("Rake assets install");
-            let rake_detect =
-                crate::steps::detect_rake_tasks(section.as_ref(), &gem_list, &context, &env)?;
+        build_output = {
+            let (bullet, rake_detect) = crate::steps::detect_rake_tasks(
+                build_output.bullet("Rake assets install"),
+                &gem_list,
+                &context,
+                &env,
+            )?;
 
             if let Some(rake_detect) = rake_detect {
-                crate::steps::rake_assets_install(section.as_ref(), &context, &env, &rake_detect)?;
+                crate::steps::rake_assets_install(bullet, &context, &env, &rake_detect)?
+            } else {
+                bullet
             }
-
-            section.end_section()
+            .done()
         };
-        logger.finish_logging();
-        warn_later.warn_now();
+        build_output.done();
 
         if let Some(default_process) = default_process {
             BuildResultBuilder::new()