+++ commons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for commons features
 
+## 2024-08-15
+
+### Changed
+
+- Deprecate `AppCacheCollection` (https://github.com/heroku/buildpacks-ruby/pull/334)
+
 ## 1.0.0
 
 ### Changed